Some minor improvements to README
I thought it would be a good idea to write a 'bit' more words about the library. So I've added goals in case some people want to ask `Why are you not using Int instead of Natural?`. I've added a quote (it was hard to find something relevant but I hope this one is nice). And I expanded library features into enumerated lists to README will look more solid.